migrate timeseries-tests to own test file
